added access token and refresh token feature

+++ server/src/models/rel_dosen_kelas.js
@@ -1,7 +1,7 @@
 module.exports = (sequelize, DataTypes) => {
     const Rel_Dosen_Kelas = sequelize.define('Rel_dosen_kelas', {
         id: {
-            type: DataTypes.INTEGER(20).UNSIGNED,
+            type: DataTypes.INTEGER(11).UNSIGNED,
             allowNull: false,
             primaryKey: true,
             autoIncrement: true
@@ -11,7 +11,7 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: false
         },
         kode_seksi: { 
-            type: DataTypes.INTEGER(20).UNSIGNED,
+            type: DataTypes.INTEGER(11).UNSIGNED,
             allowNull: false
         }
     }, {
