+++ server/src/models/rel_kelas_paketsoal.js
@@ -1,13 +1,13 @@
 module.exports = (sequelize, DataTypes) => {
     const Rel_Kelas_PaketSoal = sequelize.define('Rel_kelas_paketsoal', {
         id: {
-            type: DataTypes.INTEGER(20).UNSIGNED,
+            type: DataTypes.INTEGER(11).UNSIGNED,
             allowNull: false,
             primaryKey: true,
             autoIncrement: true
         },
         kode_seksi: { 
-            type: DataTypes.INTEGER(20).UNSIGNED,
+            type: DataTypes.INTEGER(11).UNSIGNED,
             allowNull: false
         },
         kode_paket: {
